+++ opentelemetry-exporter-gcp-logging/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-Added support for when `bytes` or `list['bytes']` is in `LogRecord.body` and
-body is of type Mapping. Update opentelemetry-api/sdk dependencies to 1.3.
+- Added support for when a `Mapping[str, bytes]` or `Mapping[str, List[bytes]]` is in `LogRecord.body`.
+- Added support for when a `Mapping[str, List[Mapping]]` is in `LogRecord.body`.
+- Do not call `logging.warning` when `LogRecord.body` is of None type, instead leave `LogEntry.payload` empty.
+- Update opentelemetry-api/sdk dependencies to 1.3.
 
 The suffix part of `LogEntry.log_name` will be the `LogRecord.event_name` when
 that is present and the `gcp.log_name` attribute is not.